+++ DevJourney/force-app/main/default/lwc/accountsComponent/accountsComponent.js
@@ -1,5 +1,6 @@
 import { LightningElement } from 'lwc';
 import { ShowToastEvent } from 'lightning/platformShowToastEvent';
+import { sortData } from 'c/utilityComponent';
 
 import getMyAccounts from '@salesforce/apex/AccountsComponentController.getMyAccounts';
 import getRecentlyViewedAccounts from '@salesforce/apex/AccountsComponentController.getRecentlyViewedAccounts';
@@ -55,14 +56,14 @@ export default class AccountsComponent extends LightningElement {
     handleMyAccountsSort(event) {
         this.myAccountsSortedBy = event.detail.fieldName;
         this.myAccountsSortDirection = event.detail.sortDirection;
-        this.myAccountsFullData = this.sortData(this.myAccountsFullData, this.myAccountsSortedBy, this.myAccountsSortDirection);
+        this.myAccountsFullData = sortData(this.myAccountsFullData, this.myAccountsSortedBy, this.myAccountsSortDirection);
         this.myAccountsData = this.myAccountsFullData.slice(0, this.myAccountsRecordCount);
     }
 
     handleRecentlyViewedAccountsSort(event) {
         this.recentlyViewedAccountsSortedBy = event.detail.fieldName;
         this.recentlyViewedAccountsSortDirection = event.detail.sortDirection;
-        this.recentlyViewedAccountsFullData = this.sortData(this.recentlyViewedAccountsFullData, this.recentlyViewedAccountsSortedBy, this.recentlyViewedAccountsSortDirection);
+        this.recentlyViewedAccountsFullData = sortData(this.recentlyViewedAccountsFullData, this.recentlyViewedAccountsSortedBy, this.recentlyViewedAccountsSortDirection);
         this.recentlyViewedAccountsData = this.recentlyViewedAccountsFullData.slice(0, this.recentlyViewedAccountRecordCount);
     }
 
@@ -83,30 +84,6 @@ export default class AccountsComponent extends LightningElement {
     /*
      * @description     Reusable code.
      */
-    sortData(data, fieldName, direction) {
-        let sortedData = [...data];
-        let getFieldValue = (record) => {
-            return record[fieldName] ? record[fieldName].toString().toLowerCase() : '';
-        };
-        let isReverse = direction === 'asc' ? 1 : -1;
-
-        sortedData.sort((x, y) => {
-            let xValue = getFieldValue(x);
-            let yValue = getFieldValue(y);
-
-            if ((xValue === '' && yValue !== '') || (xValue === null && yValue !== null)) {
-                return 1;
-            } else if ((xValue !== '' && yValue === '') || (xValue !== null && yValue === null)) {
-                return -1;
-            } else if ((xValue === '' && yValue === '') || (xValue === null && yValue === null)) {
-                return 0;
-            } else {
-                return isReverse * ((xValue > yValue) - (yValue > xValue));
-            }
-        });
-        return sortedData;
-    }
-
     loadMyAccounts() {
         this.isMyAccountsLoading = true;
         getMyAccounts()
